helm-release.yml: disable chart changes verification
The current scripts rely on Git tags, and it was taken from a repository that was dedicated exclusively to a helm chart. This won't work on this directory because it checks that tag v1.0.0 does not exist before releasing.

Since this is a manual action, I'm just disabling this check and let automatize/improve it for the future.