Treat go binaries without offsets as generic (1.9 branch patch) (#1477)
* Treat go binaries without offsets as generic

There is a small corner case in which we detect no go offsets in a
target executable, but if that executable happens to have a ".gosyms"
section, we will end up classifying it as a Go executable, causing
issues. We attach an error in this case to have the attacher treat it as
a generic executable.

* Fixes panic on Go instrumentation

On ELF parse error, we return the list of missing fields, assuming that
we would stop the instrumentation. There is an edge case error on which
the missing fields set is empty so Beyla keeps with the instrumentation,
but the returned list of fields is nil.

* Treat Go proxies as generic binaries

In the case in which we have found offsets, but these are related to a
go proxy, we must set the error so that the attacher will treat this as
a regular binary.

---------

Co-authored-by: Mario Macias <mario.macias@grafana.com>