http-spans: Align to the latest semantic conventions
I am not a huge fan of this PR, however, this is trying to kind of
follow the semconv spec.

1. We are adding a bunch of opt-in attributes. But the spec says
not to add them unless there is an explicity opt-in mechanism. I
removed them in this PR, but maybe we should hold-off until we
have a mechanism to configure the opt-in.
1. We need to add a couple more recommended attributes. The span
structure needs to change for it. Maybe I will tackle it in a
future PR

I have taken inspiration from https://github.com/open-telemetry/opentelemetry-go-contrib/blob/main/instrumentation/net/http/otelhttp/internal/semconv/httpconv.go
However, even it sometimes includes opt-in attributes by default.

Tests pending, will add them once there is consensus that we want
this change.

Signed-off-by: Goutham <gouthamve@gmail.com>